+++ infra/src/types/expression/scalar.rs
