+++ infra/src/types/memo/implementation.rs
@@ -1,6 +1,6 @@
-use crate::expression::{relational::logical::LogicalExpr, Expr};
-use crate::plan::partial_logical_plan::PartialLogicalPlan;
-use crate::plan::partial_physical_plan::PartialPhysicalPlan;
+use crate::types::expression::{relational::logical::LogicalExpr, Expr};
+use crate::types::plan::partial_logical_plan::PartialLogicalPlan;
+use crate::types::plan::partial_physical_plan::PartialPhysicalPlan;
 use crate::{ExprId, GroupId};
 
 use super::Memo;
