+++ infra/src/types/plan/logical_plan.rs
@@ -1,30 +1,25 @@
-use crate::operator::{logical::LogicalOperator, ScalarOperator};
-use crate::GroupId;
+use crate::types::operator::{logical::LogicalOperator, ScalarOperator};
 use std::sync::Arc;
 
-/// A representation of a partially materialized logical query plan DAG.
+/// A representation of a logical query plan DAG (directed acyclic graph).
 ///
-/// Similar to a [`LogicalPlan`], the `PartialLogicalPlan` DAG consists of both logical and scalar
-/// operator nodes, but it can also have unmaterialized [`GroupId`]s representing entire memo groups
-/// as children.
+/// The `LogicalPlan` DAG consists of [`LogicalOperator<LogicalLink>`] and
+/// [`ScalarOperator<ScalarLink>`] nodes, where the generic links denote the kinds of children each
+/// of those operators are allowed to have. For example, logical operators are allowed to have both
+/// logical and scalar operators as children, but scalar operators are only allowed to have other
+/// scalar operators as children.
 ///
-/// Note that while logical nodes can have both logical and scalar nodes as children, scalar nodes can only have other scalar
-/// nodes as children.
-///
-/// Note that the root of the plan _cannot_ be a scalar operator (and thus can only be a logical
+/// The root of the plan DAG _cannot_ be a scalar operator (and thus for now can only be a logical
 /// operator).
-///
-/// [`LogicalPlan`]: crate::plan::logical_plan::LogicalPlan
 #[derive(Clone)]
-pub enum PartialLogicalPlan {
+pub enum LogicalPlan {
     LogicalRoot(Arc<LogicalOperator<LogicalLink>>),
 }
 
-/// A link in a [`PartialLogicalPlan`] to a node.
+/// A link in a [`LogicalPlan`] to a node.
 ///
-/// A `LogicalLink` can be one of three things: it can be a `LogicalNode` that points to a
-/// `LogicalOperator`, it can be a `ScalarNode` that points to a `ScalarOperator`, or it can be a
-/// [`GroupId`], denoting the unamterialized part of the plan (thus the name `PartialLogicalPlan`).
+/// A `LogicalLink` can either be a `LogicalNode` that points to a `LogicalOperator` or a
+/// `ScalarNode` that points to a `ScalarOperator`.
 ///
 /// Note that this `LogicalLink` is _**different**_ from the `LogicalLink`s defined in the sibling
 /// modules:
@@ -49,20 +44,16 @@ pub enum PartialLogicalPlan {
 pub enum LogicalLink {
     LogicalNode(Arc<LogicalOperator<LogicalLink>>),
     ScalarNode(Arc<ScalarOperator<ScalarLink>>),
-    Group(GroupId),
 }
 
-/// A link in a [`LogicalPlan`] to a scalar node. A `ScalarLink` can be either a `ScalarNode` that
-/// points to a `ScalarOperator` or a `GroupId`.
+/// A link in a [`LogicalPlan`] to a scalar node. A `ScalarLink` can only be a `ScalarNode` that
+/// points to a `ScalarOperator` (for now).
 ///
 /// Note that this `ScalarLink` is _**different**_ from the `ScalarLink`s defined in the sibling
 /// modules.
 ///
 /// TODO Add detailed docs here.
-///
-/// [`LogicalPlan`]: crate::plan::logical_plan::LogicalPlan
 #[derive(Clone)]
 pub enum ScalarLink {
     ScalarNode(Arc<ScalarOperator<ScalarLink>>),
-    Group(GroupId),
 }
