+++ infra/src/types/plan/partial_physical_plan.rs
@@ -1,4 +1,4 @@
-use crate::operator::{logical::LogicalOperator, physical::PhysicalOperator, ScalarOperator};
+use crate::types::operator::{logical::LogicalOperator, physical::PhysicalOperator, ScalarOperator};
 use crate::GroupId;
 use std::sync::Arc;
